add MIN_MESSAGE_SIZE (defaults to 1KB) #17

+++ tests/tests.rb
@@ -1,21 +1,24 @@
 ENV['RACK_ENV'] = 'test'
 ENV['CALLBACK_URI_ROOT'] = 'http://localhost:9292'
 
-DEFAULT_BID = 700000
-TEST_FILE = "test.png"
-
 require 'minitest/autorun'
 require 'rack/test'
 require 'json'
 require_relative '../main'
 
+TEST_FILE = "test.png"
+TINY_TEST_FILE = "tiny_test.txt"
+
+unless File.exists?(TEST_FILE) and File.exists?(TINY_TEST_FILE)
+  `curl -o #{TEST_FILE} https://raw.githubusercontent.com/scijs/baboon-image/master/baboon.png`
+  `echo "abcdefghijklmnopqrstuvwxyz" > #{TINY_TEST_FILE}`
+end
+
+DEFAULT_BID = File.stat(TEST_FILE).size * MIN_PER_BYTE_BID + 1
+
 class MainAppTest < Minitest::Test
   include Rack::Test::Methods 
 
-  unless File.exists?(TEST_FILE)
-    `curl -o #{TEST_FILE} https://raw.githubusercontent.com/scijs/baboon-image/master/baboon.png`
-  end
- 
   def app
     Sinatra::Application
   end
@@ -72,6 +75,13 @@ class MainAppTest < Minitest::Test
 
   def test_uploaded_file_too_large
     skip "test later"
+  end
+
+  def test_uploaded_file_too_small
+    post '/order', params={"bid" => DEFAULT_BID, "file" => Rack::Test::UploadedFile.new(TINY_TEST_FILE, "text/plain")}
+    refute last_response.ok?
+    r = JSON.parse(last_response.body)
+    assert_match "too small", r["errors"][0]
   end
   
   def test_bump
